@@ -102,19 +102,15 @@ public class AuthController : ControllerBase
 
         // Generar el token de restauración
         var resetToken = GenerateJwtToken();
-        if (resetToken.Length > 255)
-        {
-            resetToken = resetToken.Substring(0, 255);
-        }
+
         user.PasswordResetToken = resetToken;
         user.PasswordResetTokenExpiry = DateTime.Now.AddHours(24);
 
         await _dbContext.SaveChangesAsync();
 
-
         // Enviar el correo
-        var resetLink = $"?token={resetToken}";
-        await _emailService.SendPasswordResetEmail(user.Email, resetLink);
+        var resetTokenEncoded = Uri.EscapeDataString(resetToken);
+        await _emailService.SendPasswordResetEmail(user.Email, resetTokenEncoded);
 
         return Ok("El enlace de restauración ha sido enviado al correo.");
     }
